+++ newlogic_main/CHANGELOG.md
@@ -1,5 +1,11 @@
 Format in CHANGELOG.md
 
-## Version[0.01] - 2022-04-03
+## Version[0.02] - 2022-05-04
+
+- [NEW] Customize res.partner to be used for the registrant model
+- [ADD] Registrants interfaces
+
+## Version[0.01] - 2022-05-03
 
 - [START] First release
+